pyproject.toml
[project]
name = "discorp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "markdown>=3.8.2",
    "pydantic-settings>=2.10.1",
    "psycopg2-binary>=2.9.9",
    "requests>=2.32.5",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "pillow>=11.3.0",
    "django-cors-headers>=4.7.0",
    "channels>=4.3.1",
    "redis>=6.4.0",
    "channels-redis>=4.3.0",
    "stripe>=12.5.0",
    "paypalrestsdk>=1.13.3",
    "cryptography>=45.0.7",
    "cloudinary>=1.44.1",
    "django-cloudinary-storage>=0.3.0",
    "drf-yasg>=1.21.10",
    "gunicorn>=21.2.0",
]
